change update styling
