+++ src/redux/actions/stylingConfig/updateStyling.js
@@ -1,10 +1,13 @@
 import STYLING_CONFIG_UPDATE from '../../actionTypes/stylingConfig';
 
-const updateStylingConfig = (experimentId, configChange) => (dispatch) => {
+const updateStylingConfig = (experimentId, setting, configChange) => (dispatch) => {
   dispatch({
     type: STYLING_CONFIG_UPDATE,
     payload:
-      { experimentId, configChange },
+    {
+      experimentId,
+      setting: configChange,
+    },
   });
 };
