+++ src/redux/reducers/stylingConfig/initialState.js
@@ -1,9 +1,82 @@
-import dataProcessingInitialConfig from './dataProcessingConfig';
-import plotsAndTablesInitialConfig from './plotsAndTablesConfig';
+import baseState from './baseState';
+
+// Plots in Data processing
+const cellSizeAndDistribution = {
+  ...baseState,
+};
+
+const mitochondrialContent = {
+  ...baseState,
+};
+
+const readAlignment = {
+  ...baseState,
+};
+
+const classifier = {
+  ...baseState,
+};
+
+const numberOfGenesVsNumberOfUMIs = {
+  ...baseState,
+};
+
+const doubletScores = {
+  ...baseState,
+};
+
+const configureEmbedding = {
+  ...baseState,
+};
+
+// Plots in Plots and Tables
+const embeddingCategoricalMain = {
+  ...baseState,
+};
+
+const embeddingContinuousMain = {
+  ...baseState,
+};
+
+const heatMap = {
+  ...baseState,
+};
+
+const frequencyPlot = {
+  ...baseState,
+};
+
+const volcanoPlot = {
+  ...baseState,
+};
 
 const initialState = {
-  ...dataProcessingInitialConfig,
-  ...plotsAndTablesInitialConfig,
+  cellSizeAndDistribution,
+  mitochondrialContent,
+  readAlignment,
+  classifier,
+  numberOfGenesVsNumberOfUMIs,
+  doubletScores,
+  configureEmbedding,
+  embeddingContinuousMain,
+  embeddingCategoricalMain,
+  heatMap,
+  volcanoPlot,
+  frequencyPlot,
 };
 
 export default initialState;
+export {
+  cellSizeAndDistribution,
+  mitochondrialContent,
+  readAlignment,
+  classifier,
+  numberOfGenesVsNumberOfUMIs,
+  doubletScores,
+  configureEmbedding,
+  embeddingContinuousMain,
+  embeddingCategoricalMain,
+  heatMap,
+  volcanoPlot,
+  frequencyPlot,
+};
